requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
numpy
pandas
scikit-learn
joblib
xgboost
selenium
undetected-chromedriver
supabase
